% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antilogit_fn.R
\name{antilogit_fn}
\alias{antilogit_fn}
\title{FUNCTION_TITLE}
\usage{
antilogit_fn(antiy, y_min, y_max, epsilon = 1e-05)
}
\arguments{
\item{antiy}{PARAM_DESCRIPTION}

\item{y_min}{PARAM_DESCRIPTION}

\item{y_max}{PARAM_DESCRIPTION}

\item{epsilon}{PARAM_DESCRIPTION, Default: 1e-05}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
