% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_forecast.R
\name{arima_forecast}
\alias{arima_forecast}
\title{Forecasting a vector of data with an ARIMA process}
\usage{
arima_forecast(data, order = c(1, 0, 0), start, end_data, end_forecast)
}
\arguments{
\item{data}{vector of data}

\item{order}{ARIMA order (defaults to an AR(1) process)}

\item{start}{first year of data available}

\item{end_data}{last year of data available}

\item{end_forecast}{last year to forecast to}
}
\value{
a vector of data with length \code{end_forecast - start} which has the forecasted data appended
}
\description{
This function will take in a vector of data and forecast using an ARIMA process as specified
}
