% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_units_of_draws.R
\name{change_units_of_draws}
\alias{change_units_of_draws}
\title{Change units of draws}
\usage{
change_units_of_draws(draws, idvar, input_data, column, op = c("add",
  "subtract", "multiply", "divide"))
}
\arguments{
\item{draws}{data.table with draws}

\item{idvar}{variables identifying the panel (country and time for example)}

\item{input_data}{data.table which has the column we'll use to do the operation with}

\item{column}{column to do the operation with}

\item{op}{unary operation to run, one of: c("add", "subtract", "multiply", "divide")}
}
\value{
a data.table with the same structure as \code{draws}, but with the operation applied
}
\description{
Run unary operations on draws
}
