% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chaos_collector_check.R
\name{chaos_collector_check}
\alias{chaos_collector_check}
\title{FUNCTION_TITLE}
\usage{
chaos_collector_check(root_fold, oos_years, chaos, scenario = F)
}
\arguments{
\item{root_fold}{PARAM_DESCRIPTION}

\item{oos_years}{PARAM_DESCRIPTION}

\item{chaos}{PARAM_DESCRIPTION}

\item{scenario}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
