% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chaos_draw_distro.R
\name{chaos_draw_distro}
\alias{chaos_draw_distro}
\title{FUNCTION_TITLE}
\usage{
chaos_draw_distro(data, chaos_pc = 10, iso_portion = 1,
  metadata = ensemble_metadata, region_portion = 0,
  super_region_portion = 0, global_portion = 0, N_draws = 1000)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{chaos_pc}{PARAM_DESCRIPTION, Default: 10}

\item{iso_portion}{PARAM_DESCRIPTION, Default: 1}

\item{metadata}{PARAM_DESCRIPTION, Default: ensemble_metadata}

\item{region_portion}{PARAM_DESCRIPTION, Default: 0}

\item{super_region_portion}{PARAM_DESCRIPTION, Default: 0}

\item{global_portion}{PARAM_DESCRIPTION, Default: 0}

\item{N_draws}{PARAM_DESCRIPTION, Default: 1000}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
