% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_by_country.R
\name{copula_by_country}
\alias{copula_by_country}
\title{FUNCTION_TITLE}
\usage{
copula_by_country(draws_array, iso, input_corr, metadata_list, AR1 = T)
}
\arguments{
\item{draws_array}{PARAM_DESCRIPTION}

\item{iso}{PARAM_DESCRIPTION}

\item{input_corr}{PARAM_DESCRIPTION}

\item{metadata_list}{PARAM_DESCRIPTION}

\item{AR1}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
