% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_draws_1D.R
\name{correlate_draws_1D}
\alias{correlate_draws_1D}
\title{FUNCTION_TITLE}
\usage{
correlate_draws_1D(draws = chaos_draws, stats = chaos_stats,
  metadata_list = metadata_list, AR1 = T)
}
\arguments{
\item{draws}{PARAM_DESCRIPTION, Default: chaos_draws}

\item{stats}{PARAM_DESCRIPTION, Default: chaos_stats}

\item{metadata_list}{PARAM_DESCRIPTION, Default: metadata_list}

\item{AR1}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
