% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_looper.R
\name{country_looper}
\alias{country_looper}
\title{FUNCTION_TITLE}
\usage{
country_looper(oos_input, years, draw_type = "diff", input_rmse_data,
  draws_data)
}
\arguments{
\item{oos_input}{PARAM_DESCRIPTION}

\item{years}{PARAM_DESCRIPTION}

\item{draw_type}{PARAM_DESCRIPTION, Default: 'diff'}

\item{input_rmse_data}{PARAM_DESCRIPTION}

\item{draws_data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
