% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_diag_data.R
\name{create_diag_data}
\alias{create_diag_data}
\title{FUNCTION_TITLE}
\usage{
create_diag_data(new_model_data, baseline_model, scatter_yrs, aroc_yrs,
  terminal_yr)
}
\arguments{
\item{new_model_data}{PARAM_DESCRIPTION}

\item{baseline_model}{PARAM_DESCRIPTION}

\item{scatter_yrs}{PARAM_DESCRIPTION}

\item{aroc_yrs}{PARAM_DESCRIPTION}

\item{terminal_yr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
