% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ensemble_grid.R
\name{create_ensemble_grid}
\alias{create_ensemble_grid}
\title{FUNCTION_TITLE}
\usage{
create_ensemble_grid(yvar, xvar, re_coef = NULL, ar = c(0),
  ar_mod = c(2), ma = c(0), ma_mod = c(2), weight_decays = c(0),
  global_int = 1, country_int = 1, country_int_dist = 1,
  fdiff = c(1), conv = c(0, 1), scaled_lev_conv = c(0),
  ar_constrain = 0)
}
\arguments{
\item{yvar}{PARAM_DESCRIPTION}

\item{xvar}{PARAM_DESCRIPTION}

\item{re_coef}{PARAM_DESCRIPTION, Default: NULL}

\item{ar}{PARAM_DESCRIPTION, Default: c(0)}

\item{ar_mod}{PARAM_DESCRIPTION, Default: c(2)}

\item{ma}{PARAM_DESCRIPTION, Default: c(0)}

\item{ma_mod}{PARAM_DESCRIPTION, Default: c(2)}

\item{weight_decays}{PARAM_DESCRIPTION, Default: c(0)}

\item{global_int}{PARAM_DESCRIPTION, Default: 1}

\item{country_int}{PARAM_DESCRIPTION, Default: 1}

\item{country_int_dist}{PARAM_DESCRIPTION, Default: 1}

\item{fdiff}{PARAM_DESCRIPTION, Default: c(1)}

\item{conv}{PARAM_DESCRIPTION, Default: c(0, 1)}

\item{scaled_lev_conv}{PARAM_DESCRIPTION, Default: c(0)}

\item{ar_constrain}{PARAM_DESCRIPTION, Default: 0}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
