% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_residuals.R
\name{create_residuals}
\alias{create_residuals}
\title{FUNCTION_TITLE}
\usage{
create_residuals(tmb_output_obj, tmb_data_param)
}
\arguments{
\item{tmb_output_obj}{PARAM_DESCRIPTION}

\item{tmb_data_param}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
