% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stats.R
\name{create_stats}
\alias{create_stats}
\title{FUNCTION_TITLE}
\usage{
create_stats(data, panelvar, variable)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{panelvar}{PARAM_DESCRIPTION}

\item{variable}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
