% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoubleMAD.R
\name{DoubleMAD}
\alias{DoubleMAD}
\title{FUNCTION_TITLE}
\usage{
DoubleMAD(x, zero.mad.action = "warn")
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{zero.mad.action}{PARAM_DESCRIPTION, Default: 'warn'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
