% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw1Dcopula.R
\name{draw1Dcopula}
\alias{draw1Dcopula}
\title{FUNCTION_TITLE}
\usage{
draw1Dcopula(X, corr_mat, print = FALSE)
}
\arguments{
\item{X}{PARAM_DESCRIPTION}

\item{corr_mat}{PARAM_DESCRIPTION}

\item{print}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
