% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw2Dcopula.R
\name{draw2Dcopula}
\alias{draw2Dcopula}
\title{FUNCTION_TITLE}
\usage{
draw2Dcopula(X, cor_mat)
}
\arguments{
\item{X}{PARAM_DESCRIPTION}

\item{cor_mat}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
