% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_run_check.R
\name{draw_run_check}
\alias{draw_run_check}
\title{FUNCTION_TITLE}
\usage{
draw_run_check(root_fold, array_grid)
}
\arguments{
\item{root_fold}{PARAM_DESCRIPTION}

\item{array_grid}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
