% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_switch.R
\name{duplicate_switch}
\alias{duplicate_switch}
\title{FUNCTION_TITLE}
\usage{
duplicate_switch(data, param_name, param_domain)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{param_name}{PARAM_DESCRIPTION}

\item{param_domain}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
