% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_cum_sum.R
\name{final_cum_sum}
\alias{final_cum_sum}
\title{FUNCTION_TITLE}
\usage{
final_cum_sum(root_fold, chaos, scenario = F, oos_years, N_draws,
  rev_trans, pop_data = NULL, pop_var = "total_pop",
  pop_action = NULL, hack_drop_NAs = F, hack_zero_bads = F)
}
\arguments{
\item{root_fold}{PARAM_DESCRIPTION}

\item{chaos}{PARAM_DESCRIPTION}

\item{scenario}{PARAM_DESCRIPTION, Default: F}

\item{oos_years}{PARAM_DESCRIPTION}

\item{N_draws}{PARAM_DESCRIPTION}

\item{rev_trans}{PARAM_DESCRIPTION}

\item{pop_data}{PARAM_DESCRIPTION, Default: NULL}

\item{pop_var}{PARAM_DESCRIPTION, Default: 'total_pop'}

\item{pop_action}{PARAM_DESCRIPTION, Default: NULL}

\item{hack_drop_NAs}{PARAM_DESCRIPTION, Default: F}

\item{hack_zero_bads}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
