% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_copula_draws.R
\name{finalize_copula_draws}
\alias{finalize_copula_draws}
\title{FUNCTION_TITLE}
\usage{
finalize_copula_draws(data_draws, copula_draws, metadata_list)
}
\arguments{
\item{data_draws}{PARAM_DESCRIPTION}

\item{copula_draws}{PARAM_DESCRIPTION}

\item{metadata_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
