% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_args_parser.R
\name{forecast_args_parser}
\alias{forecast_args_parser}
\title{FUNCTION_TITLE}
\usage{
forecast_args_parser()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
