% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entity_info.R
\name{get_entity_info}
\alias{get_entity_info}
\title{FUNCTION_TITLE}
\usage{
get_entity_info(variable = NA)
}
\arguments{
\item{variable}{PARAM_DESCRIPTION, Default: NA}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
