% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_info.R
\name{get_model_info}
\alias{get_model_info}
\title{FUNCTION_TITLE}
\usage{
get_model_info(entity_id = NA)
}
\arguments{
\item{entity_id}{PARAM_DESCRIPTION, Default: NA}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
