% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFModel_funks.r
\name{get_model_specs}
\alias{get_model_specs}
\title{Get model specifications from an ensemble grid}
\usage{
get_model_specs(data, model, draws = F)
}
\arguments{
\item{data}{data.table, an ensemble grid of model specification, created from `create_ensemble_grid`}

\item{model}{int > 0, model IF}

\item{draws}{int > 0, number of draws to assign to this model}
}
\value{
a list with the following keys:
xvar, yvar, re_coef,
ar, ar_mod, ma, ma_mod, weight_decay,
global_int, country_int, country_int_dist,
fdiff, conv, scaled_lev_conv,
draws, model, ar_constrain
}
\description{
## This function will take in the full array grid and return a list with all the specs
}
\examples{
library("data.table")
\dontrun{
# Create the ensemble grid
ens_grid <- create_ensemble_grid(yvar = 'Y', xvar = c('X1', 'X2'), ar = 1)

# Run function
get_model_specs(ens_grid, 1)
}


}
