% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nc_value_name.R
\name{get_nc_value_name}
\alias{get_nc_value_name}
\title{Get data information from xarray file}
\usage{
get_nc_value_name(nc_file)
}
\arguments{
\item{nc_file}{filepath of the NetCDF file}
}
\value{
a vector of variable names
}
\description{
Get dimension information from xarray file
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
