% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oots.R
\name{get_oots}
\alias{get_oots}
\title{FUNCTION_TITLE}
\usage{
get_oots(entity_id, model_id, scenario = 0, year_id = c(1980:2050),
  location_id = -1)
}
\arguments{
\item{entity_id}{PARAM_DESCRIPTION}

\item{model_id}{PARAM_DESCRIPTION}

\item{scenario}{PARAM_DESCRIPTION, Default: 0}

\item{year_id}{PARAM_DESCRIPTION, Default: c(1980:2050)}

\item{location_id}{PARAM_DESCRIPTION, Default: -1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
