% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_past_corr.R
\name{get_past_corr}
\alias{get_past_corr}
\title{FUNCTION_TITLE}
\usage{
get_past_corr(data_stats, metadata_list)
}
\arguments{
\item{data_stats}{PARAM_DESCRIPTION}

\item{metadata_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
