% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_logit_trans.R
\name{inv_logit_trans}
\alias{inv_logit_trans}
\title{FUNCTION_TITLE}
\usage{
inv_logit_trans(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
