% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_hold.R
\name{job_hold}
\alias{job_hold}
\title{FUNCTION_TITLE}
\usage{
job_hold(job_name, sleeper = 10)
}
\arguments{
\item{job_name}{PARAM_DESCRIPTION}

\item{sleeper}{PARAM_DESCRIPTION, Default: 10}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
