% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_fn.R
\name{logit_fn}
\alias{logit_fn}
\title{FUNCTION_TITLE}
\usage{
logit_fn(y, y_min, y_max, epsilon = 1e-05)
}
\arguments{
\item{y}{PARAM_DESCRIPTION}

\item{y_min}{PARAM_DESCRIPTION}

\item{y_max}{PARAM_DESCRIPTION}

\item{epsilon}{PARAM_DESCRIPTION, Default: 1e-05}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
