% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_diag_plots.R
\name{make_diag_plots}
\alias{make_diag_plots}
\title{FUNCTION_TITLE}
\usage{
make_diag_plots(diag_data, out_specs = c(1, 2, 3, 4, 5),
  scenarios_in_table = c(0, -1, 1))
}
\arguments{
\item{diag_data}{PARAM_DESCRIPTION}

\item{out_specs}{choices of which diagnostics to make (see Details), Default: c(1,2,3,4,5)}

\item{scenarios_in_table}{which scenarios for table to output. Default: c(0,-1,1)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
Create the following diagnostic plots and tables:
  (1) Overlapping plots of reference by countries with a baseline model
  (2) Scenarios plots of new model
  (3) Table with mean, upper and lower
  (4) Scatter of data at specific years against baseline
  (5) Scatter of AROC between past and future
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
