% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_forecast.R
\name{make_forecast}
\alias{make_forecast}
\title{FUNCTION_TITLE}
\usage{
make_forecast(mean_est_data_obj, tmb_output_object, tmb_data_param,
  add_residual = T, transform = NULL, ...)
}
\arguments{
\item{mean_est_data_obj}{PARAM_DESCRIPTION}

\item{tmb_output_object}{PARAM_DESCRIPTION}

\item{tmb_data_param}{PARAM_DESCRIPTION}

\item{add_residual}{PARAM_DESCRIPTION, Default: T}

\item{transform}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
