% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_postfiles_compile.R
\name{mean_postfiles_compile}
\alias{mean_postfiles_compile}
\title{FUNCTION_TITLE}
\usage{
mean_postfiles_compile(root_fold, passed_models, mc.cores)
}
\arguments{
\item{root_fold}{PARAM_DESCRIPTION}

\item{passed_models}{PARAM_DESCRIPTION}

\item{mc.cores}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
