% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_run_check.R
\name{mean_run_check}
\alias{mean_run_check}
\title{FUNCTION_TITLE}
\usage{
mean_run_check(root_fold, mean_array_grid, return_fail = F)
}
\arguments{
\item{root_fold}{PARAM_DESCRIPTION}

\item{mean_array_grid}{PARAM_DESCRIPTION}

\item{return_fail}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
