% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_scenario_model.R
\name{mean_scenario_model}
\alias{mean_scenario_model}
\title{FUNCTION_TITLE}
\usage{
mean_scenario_model(data, metadata_list, aroc_years, transform = "log")
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{metadata_list}{PARAM_DESCRIPTION}

\item{aroc_years}{PARAM_DESCRIPTION}

\item{transform}{PARAM_DESCRIPTION, Default: 'log'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
