% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postfile_maker.R
\name{postfile_maker}
\alias{postfile_maker}
\title{FUNCTION_TITLE}
\usage{
postfile_maker(tmb_output_obj, tmb_data_param, model_number)
}
\arguments{
\item{tmb_output_obj}{PARAM_DESCRIPTION}

\item{tmb_data_param}{PARAM_DESCRIPTION}

\item{model_number}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
