% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_array_for_copula.R
\name{prep_array_for_copula}
\alias{prep_array_for_copula}
\title{FUNCTION_TITLE}
\usage{
prep_array_for_copula(data_draws, metadata_list)
}
\arguments{
\item{data_draws}{PARAM_DESCRIPTION}

\item{metadata_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[reshape2]{cast}}
}
