% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data_for_forecast.R
\name{prep_data_for_forecast}
\alias{prep_data_for_forecast}
\title{FUNCTION_TITLE}
\usage{
prep_data_for_forecast(tmb_obj, tmb_data_param, tmb_params = NULL,
  draw_Xdata = NULL, draw_Ydata = NULL, verbose = F)
}
\arguments{
\item{tmb_obj}{PARAM_DESCRIPTION}

\item{tmb_data_param}{PARAM_DESCRIPTION}

\item{tmb_params}{PARAM_DESCRIPTION, Default: NULL}

\item{draw_Xdata}{PARAM_DESCRIPTION, Default: NULL}

\item{draw_Ydata}{PARAM_DESCRIPTION, Default: NULL}

\item{verbose}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
