% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_draw_output.R
\name{prep_draw_output}
\alias{prep_draw_output}
\title{FUNCTION_TITLE}
\usage{
prep_draw_output(draws, model)
}
\arguments{
\item{draws}{PARAM_DESCRIPTION}

\item{model}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
