% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_db_uploading.R
\name{prep_for_db_uploading}
\alias{prep_for_db_uploading}
\title{FUNCTION_TITLE}
\usage{
prep_for_db_uploading(root_fold, metadata_list, entity_info, model_num,
  custom_data = NULL, custom_draw_data_path = NULL)
}
\arguments{
\item{root_fold}{PARAM_DESCRIPTION}

\item{metadata_list}{PARAM_DESCRIPTION}

\item{entity_info}{PARAM_DESCRIPTION}

\item{model_num}{PARAM_DESCRIPTION}

\item{custom_data}{PARAM_DESCRIPTION, Default: NULL}

\item{custom_draw_data_path}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
