% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_model.R
\name{prep_model}
\alias{prep_model}
\title{FUNCTION_TITLE}
\usage{
prep_model(variable, comment, date = gsub("-", "", Sys.Date()),
  erase = F)
}
\arguments{
\item{variable}{PARAM_DESCRIPTION}

\item{comment}{PARAM_DESCRIPTION}

\item{date}{PARAM_DESCRIPTION, Default: gsub("-", "", Sys.Date())}

\item{erase}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
