% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepping_X_for_draw.R
\name{prepping_X_for_draw}
\alias{prepping_X_for_draw}
\title{FUNCTION_TITLE}
\usage{
prepping_X_for_draw(X_diff, X_level, tmb_data_param = data_params, d)
}
\arguments{
\item{X_diff}{PARAM_DESCRIPTION}

\item{X_level}{PARAM_DESCRIPTION}

\item{tmb_data_param}{PARAM_DESCRIPTION, Default: data_params}

\item{d}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
