% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_diag_plots.R
\name{print_diag_plots}
\alias{print_diag_plots}
\title{FUNCTION_TITLE}
\usage{
print_diag_plots(diag_plots, root_fold)
}
\arguments{
\item{diag_plots}{PARAM_DESCRIPTION}

\item{root_fold}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
