% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_coefs_mean.R
\name{prior_coefs_mean}
\alias{prior_coefs_mean}
\title{FUNCTION_TITLE}
\usage{
prior_coefs_mean(data, coef_prior_list)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{coef_prior_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
