% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_model_and_stop.R
\name{record_model_and_stop}
\alias{record_model_and_stop}
\title{FUNCTION_TITLE}
\usage{
record_model_and_stop(model, variable, comment, date = gsub("-", "",
  Sys.Date()), type)
}
\arguments{
\item{model}{PARAM_DESCRIPTION}

\item{variable}{PARAM_DESCRIPTION}

\item{comment}{PARAM_DESCRIPTION}

\item{date}{PARAM_DESCRIPTION, Default: gsub("-", "", Sys.Date())}

\item{type}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
