% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSE_generator.R
\name{RMSE_generator}
\alias{RMSE_generator}
\title{FUNCTION_TITLE}
\usage{
RMSE_generator(mean_data, OOS_data, panelvar, yearvar, depvar_dt,
  oos_start, in_sample_end, region_metrics = F, region_data = NA)
}
\arguments{
\item{mean_data}{PARAM_DESCRIPTION}

\item{OOS_data}{PARAM_DESCRIPTION}

\item{panelvar}{PARAM_DESCRIPTION}

\item{yearvar}{PARAM_DESCRIPTION}

\item{depvar_dt}{PARAM_DESCRIPTION}

\item{oos_start}{PARAM_DESCRIPTION}

\item{in_sample_end}{PARAM_DESCRIPTION}

\item{region_metrics}{PARAM_DESCRIPTION, Default: F}

\item{region_data}{PARAM_DESCRIPTION, Default: NA}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
