% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFModel_funks.r
\name{rmvnorm_prec}
\alias{rmvnorm_prec}
\title{Take multivariate normal draws given a mean vector and precision matrix}
\usage{
rmvnorm_prec(mu = NULL, prec, n.sims)
}
\arguments{
\item{mu}{vector of parameter means}

\item{prec}{joint precision matrix}

\item{n.sims}{number of draws}

\item{ncores}{number of cores to thread on}
}
\value{
n.sims by length(mu) matrix of parameter draws
}
\description{
Take multivariate normal draws given a mean vector and precision matrix
}
