% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{FUNCTION_TITLE}
\usage{
run_model(data_params, model, verbose = F)
}
\arguments{
\item{data_params}{PARAM_DESCRIPTION}

\item{model}{PARAM_DESCRIPTION}

\item{verbose}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[Matrix]{solve-methods}}
 \code{\link[testthat]{equality-expectations}}
}
