% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_uncertainty.R
\name{scenario_uncertainty}
\alias{scenario_uncertainty}
\title{FUNCTION_TITLE}
\usage{
scenario_uncertainty(mean_data, draws_data, metadata_list, transform,
  rev_trans, reverse_scenario = F)
}
\arguments{
\item{mean_data}{PARAM_DESCRIPTION}

\item{draws_data}{PARAM_DESCRIPTION}

\item{metadata_list}{PARAM_DESCRIPTION}

\item{transform}{PARAM_DESCRIPTION, Default: 'log'}

\item{reverse_scenario}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
