% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa_bounds.R
\name{sfa_bounds}
\alias{sfa_bounds}
\title{FUNCTION_TITLE}
\usage{
sfa_bounds(data, depvar, transform, panelvar, yearvar = "year",
  plot = F)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{depvar}{PARAM_DESCRIPTION}

\item{transform}{PARAM_DESCRIPTION}

\item{panelvar}{PARAM_DESCRIPTION}

\item{yearvar}{PARAM_DESCRIPTION, Default: 'year'}

\item{plot}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[arm]{invlogit}}
 \code{\link[frontier]{sfa}}
 \code{\link[ggplot2]{ggplot}}
}
