% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_rw_forecast.R
\name{simulate_rw_forecast}
\alias{simulate_rw_forecast}
\title{FUNCTION_TITLE}
\usage{
simulate_rw_forecast(tmb_output_object, tmb_data_params,
  var_metric = "mad", cap_variance = T, var_threshold = NULL)
}
\arguments{
\item{tmb_output_object}{PARAM_DESCRIPTION}

\item{tmb_data_params}{PARAM_DESCRIPTION}

\item{var_metric}{PARAM_DESCRIPTION, Default: 'mad'}

\item{cap_variance}{PARAM_DESCRIPTION, Default: T}

\item{var_threshold}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
