% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_bounds.R
\name{test_bounds}
\alias{test_bounds}
\title{FUNCTION_TITLE}
\usage{
test_bounds(data, depvar, predict_years, rev_trans, panelvar, yearvar,
  sfa_output, threshold = 0)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{depvar}{PARAM_DESCRIPTION}

\item{predict_years}{PARAM_DESCRIPTION}

\item{rev_trans}{PARAM_DESCRIPTION}

\item{panelvar}{PARAM_DESCRIPTION}

\item{yearvar}{PARAM_DESCRIPTION}

\item{sfa_output}{PARAM_DESCRIPTION}

\item{threshold}{PARAM_DESCRIPTION, Default: 0}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
