% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_sig.R
\name{test_sig}
\alias{test_sig}
\title{FUNCTION_TITLE}
\usage{
test_sig(tmb_output_object, alpha = 0.1, intercept_test = T,
  rigorous = F)
}
\arguments{
\item{tmb_output_object}{PARAM_DESCRIPTION}

\item{alpha}{PARAM_DESCRIPTION, Default: 0.1}

\item{intercept_test}{PARAM_DESCRIPTION, Default: T}

\item{rigorous}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
