% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_data_to_db.R
\name{upload_data_to_db}
\alias{upload_data_to_db}
\title{FUNCTION_TITLE}
\usage{
upload_data_to_db(upload_prepz)
}
\arguments{
\item{upload_prepz}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
