% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_units_of_draws.R
\name{change_units_of_draws}
\alias{change_units_of_draws}
\title{FUNCTION_TITLE}
\usage{
change_units_of_draws(draws, idvar, input_data, column, op = c("add",
  "subtract", "multiply", "divide"))
}
\arguments{
\item{draws}{PARAM_DESCRIPTION}

\item{idvar}{PARAM_DESCRIPTION}

\item{input_data}{PARAM_DESCRIPTION}

\item{column}{PARAM_DESCRIPTION}

\item{op}{PARAM_DESCRIPTION, Default: c("add", "subtract", "multiply", "divide")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
