% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antilogit_fn.R, R/logit_fn.R
\name{antilogit_fn}
\alias{antilogit_fn}
\alias{logit_fn}
\title{Bounded logit transformation package}
\usage{
antilogit_fn(y, y_min, y_max, epsilon = 1e-05)

logit_fn(y, y_min, y_max, epsilon = 1e-05)
}
\arguments{
\item{y}{number to convert to logit space}

\item{y_min}{>= 0 lower bound of domain}

\item{y_max}{upper bound of domain}

\item{epsilon}{small offset to avoid discontinuity at 0,1, Default: 1e-05}
}
\value{
a number in logit space. antilogit_fn inverts this transformation into normal space.
}
\description{
Converts a numeric into logit space on any real bounds
}
