% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_diag_plots.R
\name{make_diag_plots}
\alias{make_diag_plots}
\title{FUNCTION_TITLE}
\usage{
make_diag_plots(diag_data)
}
\arguments{
\item{diag_data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
