% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFModel_helpers.r
\name{prep_for_db_uploading}
\alias{prep_for_db_uploading}
\title{Prep output data with the entity and model info}
\usage{
prep_for_db_uploading(root_fold, metadata_list, entity_info, model_num, custom_data = NULL, custom_draw_data_path = NULL)
}
\arguments{
\item{root_fold}{the model folder info which will pull in
a data.table with the following columns:
location_id, year_id, mean, lower, upper, scenario
usually comes from Scenario output.
If no scenario columns exist, make scenario == 0 column}

\item{model_num}{Outputs include a list with:
\itemize{
 \item \code{model_info} A dataframe to be uploaded to model_versions
 \item \code{outputs} A dataframe to be uploaded to outputs
}}
}
\description{
Prep output data with the entity and model info
}
