% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvnorm_prec.R
\name{rmvnorm_prec}
\alias{rmvnorm_prec}
\title{Take multivariate normal draws given a mean vector and precision matrix}
\usage{
rmvnorm_prec(mu = NULL, prec, n.sims)
}
\arguments{
\item{mu}{vector of parameter means}

\item{prec}{joint precision matrix}

\item{n.sims}{number of draws}
}
\value{
length(mu) by n.sims matrix of parameter draws
}
\description{
Take multivariate normal draws given a mean vector and precision matrix
}
\seealso{
\code{\link[MASS]{mvrnorm}}
 \code{\link[Matrix]{Cholesky}},\code{\link[Matrix]{solve-methods}}
}
