% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvnorm_prec.R
\name{rmvnorm_prec}
\alias{rmvnorm_prec}
\title{FUNCTION_TITLE}
\usage{
rmvnorm_prec(mu = NULL, prec, n.sims)
}
\arguments{
\item{mu}{PARAM_DESCRIPTION, Default: NULL}

\item{prec}{PARAM_DESCRIPTION}

\item{n.sims}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[MASS]{mvrnorm}}
 \code{\link[Matrix]{Cholesky}},\code{\link[Matrix]{solve-methods}}
}
