% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_maker.R
\name{stat_maker}
\alias{stat_maker}
\title{FUNCTION_TITLE}
\usage{
stat_maker(data, melt = F, merge = T, idvar = c("iso3", "year"))
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{melt}{PARAM_DESCRIPTION, Default: F}

\item{merge}{PARAM_DESCRIPTION, Default: T}

\item{idvar}{PARAM_DESCRIPTION, Default: c("iso3", "year")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
